Attention Guided/requirements.txt
# Requirements for the "attmask" environment
# Python version 3.7.6

pytorch==1.7.0
torchvision==0.8.1
timm==0.4.12
tensorboardx==2.5.1
scikit-learn==0.23.2
munkres==1.1.4
tqdm
